+++ spo-un/src/main/java/com/app/spoun/dto/ProfessorDTO.java
@@ -9,7 +9,7 @@ import lombok.ToString;
 @ToString
 public class ProfessorDTO {
 
-    private Integer id;
+    private Long id;
 
     private String username;
 
@@ -21,6 +21,6 @@ public class ProfessorDTO {
 
     private String document_number;
 
-    private Integer role_id;
+    private Long role_id;
 
 }
