+++ spo-un/src/main/java/com/app/spoun/dto/StudentDTO.java
@@ -9,7 +9,7 @@ import lombok.ToString;
 @ToString
 public class StudentDTO {
 
-    private Integer id;
+    private Long id;
 
     private String username;
 
@@ -21,8 +21,8 @@ public class StudentDTO {
 
     private String document_number;
 
-    private Integer professor_id;
+    private Long professor_id;
 
-    private Integer role_id;
+    private Long role_id;
 
 }
