+++ spo-un/src/main/java/com/app/spoun/repository/ITentativeScheduleRepository.java
@@ -12,22 +12,22 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ITentativeScheduleRepository extends JpaRepository<TentativeSchedule, Integer>{
+public interface ITentativeScheduleRepository extends JpaRepository<TentativeSchedule, Long>{
 
     Page<TentativeSchedule> findAll(Pageable page);
-    Optional<TentativeSchedule> findById(Integer id);
+    Optional<TentativeSchedule> findById(Long id);
     TentativeSchedule save(TentativeSchedule tentativeSchedule);
-    void deleteById(Integer id);
-    boolean existsById(Integer id);
+    void deleteById(Long id);
+    boolean existsById(Long id);
 
-    void deleteByAppointment_id(Integer id);
+    void deleteByAppointment_id(Long id);
 
-    List<TentativeSchedule> findByAppointment_id(Integer id);
+    List<TentativeSchedule> findByAppointment_id(Long id);
 
     @Query(value = "SELECT tentative_schedule.id, tentative_schedule.start_time, tentative_schedule.end_time, tentative_schedule.appointment_id\n" +
             "FROM ((student_appointment INNER JOIN appointment ON student_appointment.appointment_id = appointment.id) INNER JOIN\n" +
             "tentative_schedule ON tentative_schedule.appointment_id = appointment.id)\n" +
             "WHERE student_appointment.student_id = ?1 and appointment.start_time IS NULL and appointment.end_time IS NULL", nativeQuery = true)
-    List<TentativeSchedule> getStudentUnconfirmedScheduleByStudentId(Integer id);
+    List<TentativeSchedule> getStudentUnconfirmedScheduleByStudentId(Long id);
 
 }
